refactor(web): extract v4 transition business logic into a service
Keep the tRPC router as a thin adapter and colocate detection-window
helpers with the SDK Redis cache contract.
